Implement AppCommand framework
