--- src/main/java/gg/projecteden/jayce/commands/common/annotations/Choices.java
@@ -1,13 +1,0 @@
-package gg.projecteden.jayce.commands.common.annotations;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Target(ElementType.PARAMETER)
-@Retention(RetentionPolicy.RUNTIME)
-public @interface Choices {
-	Class<?> value();
-
-}
