--- src/main/java/gg/projecteden/jayce/commands/common/annotations/GuildCommand.java
@@ -1,14 +1,0 @@
-package gg.projecteden.jayce.commands.common.annotations;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Target(ElementType.TYPE)
-@Retention(RetentionPolicy.RUNTIME)
-public @interface GuildCommand {
-	String[] value() default "";
-	String[] exclude() default "";
-
-}
