--- src/main/java/gg/projecteden/jayce/commands/common/annotations/Role.java
@@ -1,13 +1,0 @@
-package gg.projecteden.jayce.commands.common.annotations;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Target({ElementType.TYPE, ElementType.METHOD, ElementType.PARAMETER})
-@Retention(RetentionPolicy.RUNTIME)
-public @interface Role {
-	String value();
-
-}
